backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
firebase-admin==6.2.0
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
python-dotenv==1.0.0
structlog==23.2.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
